backend/backend_intranet/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
reportlab==4.0.7
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
